Edit pancakes.html, add images
Added a new directory for images. Added 3 new
images in the directory, which includes
pancakes.jpg, adobo.jpg and
sunny-side-up.jpg.
Adobo.jpg and sunny-side-up.jpg is yet to be
utilized for two new recipes.
Added new information to pancakes, such as a
embedded image of pancake stack, description of
the recipe, the ingredients and the steps needed
to cook pancakes.